# Performance improvement for reducing problem data tensor [issue 1160]